+++ commfuncs.h
@@ -28,7 +28,7 @@
 
 using namespace std;
 
-#define VERSION "v0.951"
+#define VERSION "v0.952"
 
 #define UNKNOWN 0